requirements.txt
flask==3.0.2
flask-cors==4.0.0
pdfplumber==0.10.3
openai==1.12.0
reportlab==4.1.0
gTTS==2.5.1
python-dotenv==1.0.1 